+++ plugins/postcss-design-tokens/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Unreleased
 
-- Added `functionName` option to control the `design-token` function name.
+- Added `valueFunctionName` option to control the `design-token` function name.
 
 ### 1.0.0 (May 23, 2022)